+++ general/includes/additional_funcs.h
