+++ general/includes/basic_funcs.h
