+++ general/includes/util.h
