+++ general/srcs/additional_funcs.c
