+++ general/srcs/basic_funcs.c
