+++ general/srcs/util.c
